requirements.txt
opencv-python-headless
numpy
matplotlib
tensorflow
Pillow
scikit-learn
keras
streamlit